+++ CHANGELOG.md
@@ -6,9 +6,11 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 
 ## [Unreleased]
 ### Added
+- light and dark basemaps for data visualization
 - coupling of duplicated map panel controls
 - dependabot
 - key results for wind, pv ground and pv roof settings panels
+- scenario settings
 
 ### Changed
 - Adapt municipality label font size according to zoom level